# chore: update Dockerfile to use Python 3.12-bookworm and refactor layout logic to utilize ELK for improved node layout